Graphics.HsCharts.Charts

Description

A simple chart drawing module, using the Gloss drawing package.

Documentation

type Point2 = (Float, Float)

type Point3 = (Float, Float, Float)

Axes

data AxisScaleType

Constructors

Linear

Log

fixedScaleAxis

Arguments

:: AxisScaleType
The type of axis scale.

-> Float
Axis length.

-> Float
Minimum axis value.

-> Float
Maximum axis value.

-> Axis

Creates an axis with a given length and range.

autoScaleAxis

Arguments

:: AxisScaleType
The type of axis scale.

-> Float
Axis length.

-> [Float]
Points to be fitted on the axis.

-> Axis

Creates an axis with a given length, and automatically finds the range.

autoScaleAxis' :: AxisScaleType -> Float -> [Float] -> Axis

Same as autoScaleAxis, but always includes the origin.

Chart components

plotChartBackground :: Axis -> Axis -> Picture

Renders the background area of a chart.

plotAxisX :: Axis -> Axis -> Picture

Renders the x-axis of a chart.

plotAxisY :: Axis -> Axis -> Picture

Renders the y-axis of a chart.

plotAxes :: Axis -> Axis -> Picture

Renders both axes of a chart.

plotAxisScales :: Axis -> Axis -> (Float, Float) -> Picture

Renders the scale values for the chart's axes.

The step size automatically adjusts to avoid overlapping texts. If the step-size is 0, no values will be shown.

plotAxisScaleLabels :: Axis -> Axis -> [(Float, String)] -> Picture

Renders given labels along the x-axis.

plotGrid :: Axis -> Axis -> (Float, Float) -> Picture

Renders a grid on the background area of a chart.

The step size automatically adjusts to avoid the lines from getting too close to each other. If the step-size is 0, no lines will be drawn.

Chart types

plotBarChart :: Axis -> Axis -> [Point2] -> Picture

Plots a bar chart from a list of 2-dimensional values.

plotPointChart :: Axis -> Axis -> [Point2] -> Picture

Renders a scatter-plot from a list of 2-dimensional values.

plotPointChart' :: Picture -> Axis -> Axis -> [Point2] -> Picture

Renders a scatter-plot with a custom point picture.

plotLineChart :: Axis -> Axis -> [Point2] -> Picture

Plots a line chart from a list of 2-dimensional values.

plotAreaChart :: Axis -> Axis -> [Point2] -> Picture

Plots an area chart from a list of 2-dimensional values, filling the area between the x-axis and the y-values.

Note: The implementation could be optimized. Currently it creates a list of tri's, to avoid drawing issues with concave polygons, regardless of whether this is necessary or not.

plotBubbleChart :: Axis -> Axis -> Axis -> [Point3] -> Picture

Plots a bubble-chart using a list of 3-dimensional values. The z-value and the z-axis range determine the size of the bubbles.

plotBubbleChart' :: (Point3 -> Picture) -> Axis -> Axis -> Axis -> [Point3] -> Picture

Plots a bubble-chart using a custom function to draw the bubbles.

boxPlot

Arguments

:: Float
Box plot width

-> Color
Box fill color

-> Axis
Vertical axis

-> [Float]
Plot data

-> Picture

Plots a vertical boxplot from a list of 1-dimensional values.

boxPlots

Arguments

:: Axis
Horizontal axis

-> Axis
Vertical axis

-> Float
Box plot width

-> Color
Box fill color

-> [Point2]
Plot data

-> Picture

Renders vertical boxplots along the x-axis

plotPolarChart

Arguments

:: ([Point2] -> Picture)
Rendering function.

-> Axis
Angular axis. Use angularAxis function to construct an appropriate Axis.

-> Axis
Radial axis. Use radialAxis function to construct an appropriate Axis.

-> [Point2]
Plot data.

-> Picture

Renders a data series in a polar coordinate system.

angularAxis

Arguments

:: Float
Chart width.

-> Axis

Creates an Axis to be used as the angular axis in a polar chart.

radialAxis :: AxisScaleType -> Float -> Float -> Float -> Axis

Creates an Axis to be used as the radial axis in a polar chart.

plotPolarGrid :: Axis -> Axis -> (Float, Float) -> Picture

Renders a polar grid on the background area of a polar chart.