Graphics.HsCharts.Components.Axis

Description

A renderable axis, with grid lines and scales.

The step sizes automatically adjusts to avoid the lines from getting too close to each other. If the step-size is 0, no lines will be drawn.

Documentation

data Axis

data AxisPosition

Constructors

AHorizontal
Horizontal axis; its position is determined by the vertical axis.

AVertical
Vertical axis; its position is determined by the horizontal axis.

ALeft
Vertical axis on the leftmost side of the chart.

ARight
Vertical axis on the rightmost side of the chart.

ATop
Horizontal axis at the top of the chart.

ABottom
Horizontal axis at the bottom of the chart.

Constructors

axis

Arguments

:: AxisPosition
Position of the axis.

-> Float
Scale step size.

-> Float
Major grid lines step size.

-> Float
Minor grid lines step size.

-> Axis

axis'

Arguments

:: Float
Scale offset.

-> Float
Grid lines offset.

-> AxisPosition
Position of the axis.

-> Float
Scale step size.

-> Float
Major grid lines step size.

-> Float
Minor grid lines step size.

-> Axis

defaultHAxis :: Axis

The default horizontal axis, with numbers and minor grid lines step size of 1.

defaultVAxis :: Axis

The default vertical axis, with numbers and minor grid lines step size of 1.