Graphics.HsCharts.Components.Grid

Description

Renders a grid on the background area of a chart.

The step size automatically adjusts to avoid the lines from getting too close to each other. If the step-size is 0, no lines will be drawn.

Documentation

data Grid

Constructors

grid

Arguments

:: Color
Grid color.

-> Float
Horizontal grid size.

-> Float
Vertical grid size.

-> Grid

Grid constructor

defaultGrid :: Grid

The default (light-grey) grid, with vertical and horizontal size 1.

Other functions

gridPoints

Arguments

:: AxisScale
Axis scale to generate grid points for.

-> Float
Preferred step size.

-> Float
Minimal distance in pixels between the points.

-> [(Float, Float)]
A list of grid points (axis value, scaled value).

Generates a list of points on the grid.

gridPoints'

Arguments

:: Float
Start offset

-> AxisScale
Axis scale to generate grid points for.

-> Float
Preferred step size.

-> Float
Minimal distance in pixels between the points.

-> [(Float, Float)]
A list of grid points (axis value, scaled value).