Graphics.HsCharts.Plots.AreaChart

Description

Renders a set of points with a filled area between the horizontal axis and the y-values.

Documentation

data AreaChart

Constructors

areaChart :: Color -> [Point2] -> AreaChart

Creates an area chart with the given area color.

defaultAreaChart :: [Point2] -> AreaChart

A default area chart, with a blue, semi-transparent area color.