Graphics.HsCharts.Plots.BarChart

Description

A simple vertical bar chart.

Use the barAxis and barAxisScale functions to fit the bars nicely on the horizontal axis.

Documentation

data BarChart

Constructors

barChart :: Color -> [Point2] -> BarChart

Bar Chart constructor.

defaultBarChart :: [Point2] -> BarChart

A default bar chart, with a black bar color.

Axes

barAxisScale

Arguments

:: Float
The length of the axis.

-> [Float]
The x-values to be fitted on the axis.

-> AxisScale

Creates a horizontal axis scale.

barAxis :: Axis

An axis with a 0.5 scale numbers offset.