Graphics.HsCharts.Plots.PolarChart

Documentation

data PolarChart

polarChart :: ([Point2] -> Picture) -> [Point2] -> PolarChart

Polar Chart constructor.

defaultPolarChart :: [Point2] -> PolarChart

Axes

angularAxis

Arguments

:: Float
Chart width.

-> AxisScale

Creates an Axis to be used as the angular axis in a polar chart.

radialAxis :: AxisScaleType -> Float -> Float -> Float -> AxisScale

Creates an Axis to be used as the radial axis in a polar chart.

Grid

data PolarGrid

polarGrid :: Color -> Float -> Float -> PolarGrid