Graphics.HsCharts.Shapes

Description

Various compound Gloss shapes.

Documentation

eqTrianglePath :: Float -> Path

Creates an equilateral triangle (all sides same length) of the specified size.

eqTriangleSolid :: Float -> Picture

Draws a solid equilateral triangle (all sides same length) of the specified size.

cross :: Float -> Picture

Draws a cross of the specified size.

diamondPath :: Float -> Float -> Path

Creates a diamond with the specified size and aspect.

diamondSolid :: Float -> Float -> Picture

Draws a solid diamond with the specified size and aspect.

concavePolygon :: [Point] -> Picture

Concave polygon - not implemented yet.