Graphics.HsCharts.Types

Documentation

type Point2 = (Float, Float)

type Point3 = (Float, Float, Float)

Axis Scale

data AxisScale

Constructors

AxisScale

Fields

axisLength :: Float

axisMin :: Float

axisMax :: Float

axisScaleFn :: Float -> Float

data AxisScaleType

Constructors

Linear

Log

fixedAxisScale

Arguments

:: AxisScaleType
The type of axis scale.

-> Float
Axis length.

-> Float
Minimum axis value.

-> Float
Maximum axis value.

-> AxisScale

Creates an axis with a given length and range.

autoAxisScale

Arguments

:: AxisScaleType
The type of axis scale.

-> Float
Axis length.

-> [Float]
Points to be fitted on the axis.

-> AxisScale

Creates an axis with a given length, and automatically finds the range.

autoAxisScale' :: AxisScaleType -> Float -> [Float] -> AxisScale

Same as autoScaleAxis, but always includes the origin.

getAxisZero :: AxisScale -> Float

scalePoints :: AxisScale -> AxisScale -> [Point2] -> [Point2]

Gloss

class ToPicture a where

Methods

toPicture :: a -> AxisScale -> AxisScale -> Picture

plotPictures :: [AxisScale -> AxisScale -> Picture] -> AxisScale -> AxisScale -> Picture